Refactor existing Site_Goals_Settings class from Modules/Analytics namespace to Core/User namespace.
